+++ R/RcppExports.R
@@ -493,8 +493,8 @@ json_to_acmap <- function(json) {
     .Call('_Racmacs_json_to_acmap', PACKAGE = 'Racmacs', json)
 }
 
-acmap_to_json <- function(map, version) {
-    .Call('_Racmacs_acmap_to_json', PACKAGE = 'Racmacs', map, version)
+acmap_to_json <- function(map, version, pretty) {
+    .Call('_Racmacs_acmap_to_json', PACKAGE = 'Racmacs', map, version, pretty)
 }
 
 ac_procrustes <- function(X, Xstar, translation, dilation) {