+++ src/RcppExports.cpp
@@ -1437,14 +1437,15 @@ BEGIN_RCPP
 END_RCPP
 }
 // acmap_to_json
-std::string acmap_to_json(AcMap map, std::string version);
-RcppExport SEXP _Racmacs_acmap_to_json(SEXP mapSEXP, SEXP versionSEXP) {
+std::string acmap_to_json(AcMap map, std::string version, bool pretty);
+RcppExport SEXP _Racmacs_acmap_to_json(SEXP mapSEXP, SEXP versionSEXP, SEXP prettySEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< AcMap >::type map(mapSEXP);
     Rcpp::traits::input_parameter< std::string >::type version(versionSEXP);
-    rcpp_result_gen = Rcpp::wrap(acmap_to_json(map, version));
+    Rcpp::traits::input_parameter< bool >::type pretty(prettySEXP);
+    rcpp_result_gen = Rcpp::wrap(acmap_to_json(map, version, pretty));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1663,7 +1664,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_Racmacs_titer_types_int", (DL_FUNC) &_Racmacs_titer_types_int, 1},
     {"_Racmacs_reduce_matrix_dimensions", (DL_FUNC) &_Racmacs_reduce_matrix_dimensions, 2},
     {"_Racmacs_json_to_acmap", (DL_FUNC) &_Racmacs_json_to_acmap, 1},
-    {"_Racmacs_acmap_to_json", (DL_FUNC) &_Racmacs_acmap_to_json, 2},
+    {"_Racmacs_acmap_to_json", (DL_FUNC) &_Racmacs_acmap_to_json, 3},
     {"_Racmacs_ac_procrustes", (DL_FUNC) &_Racmacs_ac_procrustes, 4},
     {"_Racmacs_ac_align_coords", (DL_FUNC) &_Racmacs_ac_align_coords, 4},
     {"_Racmacs_ac_procrustes_map_coords", (DL_FUNC) &_Racmacs_ac_procrustes_map_coords, 6},